Standardised the Identity user framework to make it more intuitive and easier to implement based upon Microsoft conventions.
