+++ Code/Tardigrade.Framework/Tardigrade.Framework.AspNet/Tardigrade.Framework.AspNet.csproj
@@ -43,7 +43,9 @@
   <ItemGroup>
     <Compile Include="Models\Identity\ApplicationUser.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Services\Identity\ApplicationSignInManager.cs" />
     <Compile Include="Services\Identity\ApplicationUserManager.cs" />
+    <Compile Include="Services\Identity\IdentityUserManager.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Tardigrade.Framework\Tardigrade.Framework.csproj">
