+++ Code/Tardigrade.Framework/Tardigrade.Framework.AspNetCore/Models/Identity/ApplicationUser.cs
@@ -1,12 +1,11 @@
 ﻿using Microsoft.AspNetCore.Identity;
-using Tardigrade.Framework.Models.Identity;
 
 namespace Tardigrade.Framework.AspNetCore.Models.Identity
 {
     /// <summary>
     /// Represents an application user based upon an ASP.NET Core Identity user.
     /// </summary>
-    public class ApplicationUser : IdentityUser, IApplicationUser
+    public class ApplicationUser : IdentityUser
     {
     }
 }
