+++ Code/Tardigrade.Framework/Tardigrade.Framework.AspNetCore/Services/Identity/ApplicationSignInManager.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using Tardigrade.Framework.AspNetCore.Models.Identity;
+
+namespace Tardigrade.Framework.AspNetCore.Services.Identity
+{
+    /// <summary>
+    /// ASP.NET Core Identity sign-in manager configured for this application.
+    /// </summary>
+    public class ApplicationSignInManager : SignInManager<ApplicationUser>
+    {
+        /// <summary>
+        /// <see cref="SignInManager{TUser}.SignInManager(UserManager{TUser}, IHttpContextAccessor, IUserClaimsPrincipalFactory{TUser}, IOptions{IdentityOptions}, ILogger{SignInManager{TUser}}, IAuthenticationSchemeProvider)"/>
+        /// </summary>
+        public ApplicationSignInManager(
+            UserManager<ApplicationUser> userManager,
+            IHttpContextAccessor contextAccessor,
+            IUserClaimsPrincipalFactory<ApplicationUser> claimsFactory,
+            IOptions<IdentityOptions> optionsAccessor,
+            ILogger<SignInManager<ApplicationUser>> logger,
+            IAuthenticationSchemeProvider schemes)
+            : base(userManager, contextAccessor, claimsFactory, optionsAccessor, logger, schemes)
+        {
+        }
+    }
+}
