+++ Code/Tardigrade.Framework/Tardigrade.Framework.EntityFramework/Tardigrade.Framework.EntityFramework.csproj
@@ -41,7 +41,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Data\ApplicationDbContext.cs" />
+    <Compile Include="Data\IdentityUserDbContext.cs" />
     <Compile Include="EntityFrameworkRepository.cs" />
     <Compile Include="EntityFrameworkUnitOfWork.cs" />
     <Compile Include="Extensions\RepositoryExtension.cs" />
@@ -50,9 +50,6 @@
   <ItemGroup>
     <PackageReference Include="EntityFramework">
       <Version>6.2.0</Version>
-    </PackageReference>
-    <PackageReference Include="Microsoft.AspNet.Identity.EntityFramework">
-      <Version>2.2.2</Version>
     </PackageReference>
   </ItemGroup>
   <ItemGroup>
