+++ Code/Tardigrade.Framework/Tardigrade.Framework.EntityFrameworkCore/Data/IdentityUserDbContext.cs
@@ -5,15 +5,15 @@ using Tardigrade.Framework.AspNetCore.Models.Identity;
 namespace Tardigrade.Framework.EntityFrameworkCore.Data
 {
     /// <summary>
-    /// Entity Framework Core database context for the application.
+    /// Entity Framework Core database context for use with ASP.NET Core Identity.
     /// </summary>
-    public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
+    public class IdentityUserDbContext : IdentityDbContext<ApplicationUser>
     {
         /// <summary>
         /// Create an instance of this class.
         /// </summary>
         /// <param name="options">Options associated with the database context.</param>
-        public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : base(options)
+        public IdentityUserDbContext(DbContextOptions<IdentityUserDbContext> options) : base(options)
         {
         }
     }
