+++ Code/Tardigrade.Framework/Tardigrade.Framework/Models/Identity/IIdentityUser.cs
@@ -3,7 +3,7 @@
     /// <summary>
     /// Represents an application user (as opposed to a user domain model).
     /// </summary>
-    public interface IApplicationUser
+    public interface IIdentityUser
     {
     }
 }
